[Performance] Improved performance by converting 4 render targets from RHI_Format_R16G16B16A16_Float to RHI_Format_R11G11B10_Float. Unnecessary alpha channel writes have also been removed from post processing passes.

+++ Runtime/Rendering/Renderer_ConstantBuffers.h
@@ -133,7 +133,7 @@ namespace Spartan
         float mat_normal_mul;
         float mat_height_mul;
 
-        float mat_id;
+        uint32_t mat_id;
         float is_transparent_pass;
         uint32_t mip_count;
         uint32_t work_group_count;
